refactor(web): extract JS results payload into payload module
Move the Js* serializable structs, JsResults::from_results, and the
build_mask_overlay/put/task_of helpers out of lib.rs into a dedicated
crates/web/src/payload.rs. This is the Results -> Serialize boundary
(~230 lines); separating it from the session/predict flow shrinks lib.rs
and groups the JS mapping in one place. JsResults and from_results are
pub(crate); the rest stay private to the module.

No behavior change.
